Remove parse_plugin_source/get_cache_path, slim plugin tests
- Remove parse_plugin_source and get_cache_path from plugin/fetch.py;
  callers should use extensions.fetch directly.
- Slim test_plugin_fetch.py from 1038 to 100 lines: keep only
  PluginFetchError wrapping, DEFAULT_CACHE_DIR, and Plugin.fetch() tests.
- Move git infrastructure tests (clone, update, checkout, locking,
  GitHelper errors, get_default_branch) to tests/sdk/git/test_cached_repo.py.
- Update test_installed_plugins.py to import from extensions.fetch.
